+++ tests/ui/container_with_and_shrink.rs
@@ -1,0 +1,9 @@
+use quickcheck_richderive::Arbitrary;
+
+#[derive(Arbitrary)]
+#[quickcheck(with = "m", shrink = "s")]
+struct S {
+  x: u8,
+}
+
+fn main() {}