+++ tests/ui/field_with_and_arbitrary.rs
@@ -1,0 +1,9 @@
+use quickcheck_richderive::Arbitrary;
+
+#[derive(Arbitrary)]
+struct S {
+  #[quickcheck(with = "m", arbitrary = "f")]
+  x: u8,
+}
+
+fn main() {}